start: panic on error instead of passing the error back to JS
There's really little point in trying to pass most
errors back to JavaScript, in fact it just makes
debugging more difficult.

The sole exception here is when we read the game data
from a slice. If the game data is invalid, it
absolutely makes sense to report that back to
JavaScript.